+++ app/src/secrets/index.rs
@@ -29,7 +29,7 @@ pub async fn index(
         .await?;
 
     let user_vault = queries::user_vaults::get()
-        .bind(&transaction, &(current_user.user_id as i32), &vault_id)
+        .bind(&transaction, &current_user.user_id, &vault_id)
         .one()
         .await?;
 
@@ -39,7 +39,7 @@ pub async fn index(
         .await?;
 
     let user = queries::users::get()
-        .bind(&transaction, &(current_user.user_id as i32))
+        .bind(&transaction, &current_user.user_id)
         .one()
         .await?;
     let initials = crate::layout::initials(&user.email, user.first_name, user.last_name);
@@ -58,7 +58,7 @@ pub async fn index(
         queries::audit::insert()
             .bind(
                 &transaction,
-                &(current_user.user_id as i32),
+                &current_user.user_id,
                 &team.id,
                 &AuditAction::AccessSecrets,
                 &AuditAccessType::Web,
